requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crispy-bootstrap5==0.4
cryptography==3.4.5
defusedxml==0.6.0
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-imageit==0.0.6
django-recaptcha==2.0.6
filetype==1.0.7
idna==2.10
oauthlib==3.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.3
